Configure next js api-routes for fe-only development
