+++ frontend/.env.development
@@ -1,0 +1,2 @@
+# set to LOCAL for frontend only (no need servers for testing) set to SERVER for using actual backend services
+BACKEND_MODE=LOCAL
